Refactored Mission
Better share of responsibilities between mission definition, mission building and its OpenMDAO wrapping
