+++ src/fastoad/models/performances/mission/mission_definition/tests/data/mission.yml
@@ -1,7 +1,9 @@
 phase_definitions:
   initial_climb:
     engine_setting: takeoff
-    polar: data:aerodynamics:aircraft:takeoff
+    polar:
+      CL: 0.0, 0.5, 1.0
+      CD: 0., 0.03, 0.12
     thrust_rate: data:mission:sizing:climb:thrust_rate
     steps:
       - segment: altitude_change
@@ -12,12 +14,15 @@ phase_definitions:
           equivalent_airspeed:
             value: constant
       - segment: speed_change
+        polar: data:aerodynamics:aircraft:takeoff
         target:
           equivalent_airspeed:
             value: 250
             unit: kn
       - segment: altitude_change
-        #        polar: data:aerodynamics:aircraft:low_speed
+        polar:
+          CL: data:aerodynamics:aircraft:takeoff:CL
+          CD: data:aerodynamics:aircraft:takeoff:CD
         target:
           altitude:
             value: 1500.
